+++ core/src/autogluon/core/models/abstract/abstract_model.py
@@ -23,7 +23,7 @@ from ._tags import _DEFAULT_TAGS
 from ... import metrics, Space
 from ...constants import AG_ARGS_FIT, BINARY, REGRESSION, QUANTILE, REFIT_FULL_SUFFIX, OBJECTIVES_TO_NORMALIZE
 from ...data.label_cleaner import LabelCleaner, LabelCleanerMulticlassToBinary
-from ...scheduler import LocalSequentialScheduler
+from ...scheduler import LocalScheduler
 from ...utils import get_cpu_count, get_pred_from_proba, normalize_pred_probas, infer_eval_metric, infer_problem_type, \
     compute_permutation_feature_importance, compute_weighted_metric
 from ...utils.exceptions import TimeLimitExceeded, NoValidFeatures, NotEnoughMemoryError
@@ -989,10 +989,11 @@ class AbstractModel:
             val_path=val_path,
         )
 
-        scheduler: LocalSequentialScheduler = scheduler_cls(model_trial, search_space=search_space, train_fn_kwargs=train_fn_kwargs, **scheduler_params)
+        scheduler: LocalScheduler = scheduler_cls(model_trial, search_space=search_space, train_fn_kwargs=train_fn_kwargs, **scheduler_params)
 
         scheduler.run()
-        scheduler.join_jobs()
+        # FIXME: shouldn't remove this
+        # scheduler.join_jobs()
 
         return self._get_hpo_results(scheduler=scheduler, scheduler_params=scheduler_params, time_start=time_start)
